buscar,deletar e alterar por ID

+++ app/Http/Controllers/ProductController.php
@@ -14,26 +14,38 @@ class ProductController extends Controller
 
     public function show($id)
     {
-        return product::find($id);
+
+        $product =product::find($id);
+
+        if(is_null($product)) return response()->json('Product not found', 404);
+
+        return $product;
     }
 
     public function store(Request $request)
     {
         $product = product::create($request->all());
+
         return $product;
     }
 
     public function update(Request $request, $id)
     {
         $product = product::find($id);
+
         $product->update($request->all());
+        
         return $product;
     }
 
     public function destroy($id)
     {
         $product = product::find($id);
+
+        if(is_null($product)) return response()->json('Product not found', 404);
+
         $product->delete();
+
         return 204;
     }
 }
